% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthRisk.R
\name{HealthRisk}
\alias{HealthRisk}
\title{Health Risk and Benefit Estimation (V 0.1.0)}
\usage{
HealthRisk(
  file_path = "HealthRisktemplate.xlsx",
  mode = "risk",
  target_type,
  target_value
)
}
\arguments{
\item{file_path}{The file path of the baseline data, including exposure concentration, 
population, baseline incidence rate, and exposure-response relationships.}

\item{mode}{The estimation type: "risk" (default) for the health risk based on current 
exposure concentrations; "benefit" for the health benefit if the air pollutant concentrations 
are reduced by a given percentage or reduced to given target concentrations.}

\item{target_type}{The type of target value: "threshold" means that the target_value is a 
threshold concentration to which the air pollutant exposure will be reduced; "percentage"
means that the target_value is a percentage by which the the air pollutant exposure will be reduced; 
This parameter is not required when "mode" is "risk".}

\item{target_value}{The target value for controlling the air pollution. It should be a 
threshold concentration to which the air pollutant exposure will be reduced if "target_type" is "threshold"; 
it should be a percentage by which the the air pollutant exposure will be reduced if "target_type" is 
"percentage"; This parameter is not required when "mode" is "risk".}
}
\value{
The health risk or benefits in the form of mortality and incidence cases.
}
\description{
Estimate the health risk based on exposure concentrations of air pollutants, 
and health benefits if target concentrations or reduced percentage is provided.
PLEASE DOWNLOAD OUR XLSX FILE TEMPLATE REQUIRED BY THIS FUNCTION: 
\url{https://raw.githubusercontent.com/liunr24/Health_risk_R_package/main/HealthRisktemplate.xlsx}
}
\examples{
result1 <- HealthRisk();
result1 <- HealthRisk(file_path = "HealthRisktemplate.xlsx",mode = "risk"); # Health risk of current exposure


result2 <- HealthRisk(file_path = "HealthRisktemplate.xlsx",mode = "benefit",
                      target_type = "percentage",target_value = rep(0.1,24)); # Health benefit from reducing exposure by 10\%
}
